[SYCL] Implement matrix extension using new unified interface (#7413)
