+++ sycl/include/sycl/ext/oneapi/matrix/matrix-unified.hpp
@@ -7,8 +7,8 @@
 // ===--------------------------------------------------------------------=== //
 
 #pragma once
+#include "matrix-intel.hpp"
 #include <sycl/ext/oneapi/matrix/matrix-tensorcores.hpp>
-
 namespace sycl {
 __SYCL_INLINE_VER_NAMESPACE(_V1) {
 namespace ext {
@@ -20,13 +20,16 @@ template <typename Group, typename T, use Use, size_t Rows, size_t Cols,
           layout Layout>
 struct joint_matrix {
 
-#if defined(__SYCL_DEVICE_ONLY__) && defined(__SPIR__)
-  // TODO: Intel case here: we use the ext_oneapi_cuda case also for the host,
-  // because the Intel SPIRV functions will not be host compilable.
-#else
+#if defined(__SYCL_DEVICE_ONLY__)
+#if defined(__NVPTX__)
   sycl::ext::oneapi::detail::joint_matrix_cuda<T, Use, Rows, Cols, Layout>
       cuda_impl;
-#endif // defined(__SYCL_DEVICE_ONLY__) && defined(__SPIR__)
+#else
+  __spv::__spirv_JointMatrixINTEL<
+      T, Rows, Cols, spv_matrix_layout_traits<Layout>::value,
+      spv_scope_traits<Group>::value, spv_matrix_use_traits<Use>::value> *spvm;
+#endif // defined(__SYCL_DEVICE_ONLY__)
+#endif
 
   joint_matrix() {
 #ifndef __SYCL_DEVICE_ONLY__
@@ -36,17 +39,74 @@ struct joint_matrix {
   }
 };
 
+#ifdef __SYCL_DEVICE_ONLY__
 template <typename Group, typename T, use Use, size_t Rows, size_t Cols,
           layout Layout>
-inline __SYCL_ALWAYS_INLINE wi_data<Group, T, Use, Rows, Cols, Layout>
+class wi_data {
+
+  joint_matrix<Group, T, Use, Rows, Cols, Layout> &jm;
+
+  wi_data(joint_matrix<Group, T, Use, Rows, Cols, Layout> &_jm) : jm(_jm){};
+
+  template <typename Grp, typename Type, use UseJm, size_t NumRows,
+            size_t NumCols, layout LayoutJm>
+  friend decltype(auto)
+  get_wi_data(Grp,
+              joint_matrix<Grp, Type, UseJm, NumRows, NumCols, LayoutJm> &);
+
+public:
+  size_t length() {
+#if defined(__NVPTX__)
+    return jm.cuda_impl.wi_marray.size();
+#else
+    return __spirv_JointMatrixWorkItemLengthINTEL(jm.spvm);
+#endif
+  };
+
+  decltype(auto) operator[](size_t i) {
+#if defined(__NVPTX__)
+    return (jm.cuda_impl.wi_marray[i]);
+#else
+    return wi_element<T, Rows, Cols, Use, Layout, Group>(jm, i);
+#endif
+  };
+};
+#else
+template <typename type, size_t size> class wi_data {
+  marray<type, size> &data;
+  wi_data(marray<type, size> &wi_marray) : data(wi_marray){};
+  template <typename Grp, typename Type, use UseJm, size_t NumRows,
+            size_t NumCols, layout LayoutJm>
+  friend decltype(auto)
+  get_wi_data(Grp,
+              joint_matrix<Grp, Type, UseJm, NumRows, NumCols, LayoutJm> &);
+
+public:
+  size_t length() { return data.size(); };
+
+  type &operator[](size_t i) { return data[i]; };
+};
+#endif
+
+template <typename Group, typename T, use Use, size_t Rows, size_t Cols,
+          layout Layout>
+inline __SYCL_ALWAYS_INLINE decltype(auto)
 get_wi_data(Group sg, joint_matrix<Group, T, Use, Rows, Cols, Layout> &jm) {
 #if defined(__SYCL_DEVICE_ONLY__)
 #if defined(__NVPTX__)
   std::ignore = sg;
   return wi_data(jm);
 #else
-  // TODO add Intel impl.
+  return wi_data<Group, T, Use, Rows, Cols, Layout>(jm);
 #endif // defined(__NVPTX__)
+#else
+  if constexpr (std::is_same_v<T, precision::tf32>) {
+    marray<float, 1> unused{};
+    return wi_data<float, 1>(unused);
+  } else {
+    marray<T, 1> unused{};
+    return wi_data<T, 1>(unused);
+  }
 #endif // defined(__SYCL_DEVICE_ONLY__)
 }
 
@@ -60,6 +120,12 @@ joint_matrix_fill(Group sg,
 #if defined(__NVPTX__)
   std::ignore = sg;
   res.cuda_impl.wi_marray = v;
+#else
+  res.spvm =
+      __spirv_CompositeConstruct<T, NumRows, NumCols,
+                                 spv_matrix_use_traits<Use>::value,
+                                 spv_matrix_layout_traits<Layout>::value>(
+          static_cast<T>(v));
 #endif // defined(__NVPTX__)
 #else
   std::ignore = sg;
@@ -88,6 +154,35 @@ inline __SYCL_ALWAYS_INLINE void joint_matrix_load(
   std::ignore = sg;
   sycl::ext::oneapi::detail::load_accumulator_cuda(res.cuda_impl, src, stride,
                                                    Layout);
+#else
+  // intel's impl
+  // matL is determined by matrix.use?
+  T *Ptr = src.get();
+  switch (Layout) {
+  default:
+    assert(false && "Invalid Memory Layout!");
+  case layout::row_major:
+    res.spvm = __spirv_JointMatrixLoadINTEL<
+        T, NumRows, NumCols, spv_matrix_use_traits<use::accumulator>::value,
+        spv_matrix_layout_traits<layout::dynamic>::value>(
+        Ptr, stride, __spv::MatrixLayout::RowMajor,
+        spv_scope_traits<Group>::value);
+    break;
+  case layout::col_major:
+    res.spvm = __spirv_JointMatrixLoadINTEL<
+        T, NumRows, NumCols, spv_matrix_use_traits<use::accumulator>::value,
+        spv_matrix_layout_traits<layout::dynamic>::value>(
+        Ptr, stride, __spv::MatrixLayout::ColumnMajor,
+        spv_scope_traits<Group>::value);
+    break;
+  case sycl::ext::intel::experimental::matrix::layout::packed:
+    res.spvm = __spirv_JointMatrixLoadINTEL<
+        T, NumRows, NumCols, spv_matrix_use_traits<use::accumulator>::value,
+        spv_matrix_layout_traits<layout::dynamic>::value>(
+        Ptr, stride, __spv::MatrixLayout::Packed,
+        spv_scope_traits<Group>::value);
+    break;
+  }
 #endif // defined(__NVPTX__)
 #else
   std::ignore = sg;
@@ -119,6 +214,14 @@ joint_matrix_load(Group sg,
   sycl::ext::oneapi::detail::load_multiplicand_cuda<S, T, NumRows, NumCols, Use,
                                                     Layout, Space>(
       res.cuda_impl, src, stride);
+#else
+  T *Ptr = src.get();
+  res.spvm =
+      __spirv_JointMatrixLoadINTEL<T, NumRows, NumCols,
+                                   spv_matrix_use_traits<Use>::value,
+                                   spv_matrix_layout_traits<Layout>::value>(
+          Ptr, stride, spv_matrix_layout_traits<Layout>::value,
+          spv_scope_traits<Group>::value);
 #endif // defined(__NVPTX__)
 #else
   std::ignore = sg;
@@ -146,6 +249,34 @@ inline __SYCL_ALWAYS_INLINE void joint_matrix_store(
   sycl::ext::oneapi::detail::joint_matrix_store_cuda<T, NumRows, NumCols,
                                                      Space>(src.cuda_impl, dst,
                                                             stride, Layout);
+#else
+  // intel's impl
+  T *Ptr = dst.get();
+  switch (Layout) {
+  default:
+    assert(false && "Invalid Memory Layout!");
+  case layout::row_major:
+    __spirv_JointMatrixStoreINTEL<
+        T, NumRows, NumCols, spv_matrix_use_traits<use::accumulator>::value,
+        spv_matrix_layout_traits<layout::dynamic>::value>(
+        Ptr, src.spvm, stride, __spv::MatrixLayout::RowMajor,
+        spv_scope_traits<Group>::value);
+    break;
+  case layout::col_major:
+    __spirv_JointMatrixStoreINTEL<
+        T, NumRows, NumCols, spv_matrix_use_traits<use::accumulator>::value,
+        spv_matrix_layout_traits<layout::dynamic>::value>(
+        Ptr, src.spvm, stride, __spv::MatrixLayout::ColumnMajor,
+        spv_scope_traits<Group>::value);
+    break;
+  case sycl::ext::intel::experimental::matrix::layout::packed:
+    __spirv_JointMatrixStoreINTEL<
+        T, NumRows, NumCols, spv_matrix_use_traits<use::accumulator>::value,
+        spv_matrix_layout_traits<layout::dynamic>::value>(
+        Ptr, src.spvm, stride, __spv::MatrixLayout::Packed,
+        spv_scope_traits<Group>::value);
+    break;
+  }
 #endif // defined(__NVPTX__)
 #else
   std::ignore = sg;
@@ -185,6 +316,21 @@ inline __SYCL_ALWAYS_INLINE
     assert(false && "Ta != Tb : In the CUDA backend joint_matrix_mad "
                     "requires that joint_matrix data types Ta and Tb match");
   }
+#else
+  joint_matrix<Group, Tc, use::accumulator, M, N, layout::dynamic> res;
+  if constexpr (std::is_same<Ta, uint16_t>::value &&
+                std::is_same<Tb, uint16_t>::value &&
+                std::is_same<Tc, float>::value)
+    res.spvm = __spirv_JointMatrixMadINTEL(A.spvm, B.spvm, C.spvm);
+  else if constexpr (std::is_unsigned<Ta>::value && std::is_unsigned<Tb>::value)
+    res.spvm = __spirv_JointMatrixUUMadINTEL(A.spvm, B.spvm, C.spvm);
+  else if constexpr (std::is_signed<Ta>::value && std::is_unsigned<Tb>::value)
+    res.spvm = __spirv_JointMatrixSUMadINTEL(A.spvm, B.spvm, C.spvm);
+  else if constexpr (std::is_unsigned<Ta>::value && std::is_signed<Tb>::value)
+    res.spvm = __spirv_JointMatrixUSMadINTEL(A.spvm, B.spvm, C.spvm);
+  else
+    res.spvm = __spirv_JointMatrixMadINTEL(A.spvm, B.spvm, C.spvm);
+  return res;
 #endif // defined(__NVPTX__)
 #else
   std::ignore = sg;
