+++ sycl/test/matrix/legacy/matrix-elemwise-ops.cpp
@@ -1,4 +1,5 @@
-// RUN: %clangxx -fsycl -O2 %s -o %t.out
+// RUN: %clangxx -fsycl -O2 -DSYCL_EXT_ONEAPI_MATRIX_VERSION=1 %s -o %t.out
+
 #include <iostream>
 #include <sycl/sycl.hpp>
 
@@ -6,26 +7,28 @@ using namespace sycl;
 using namespace sycl::ext::oneapi::experimental::matrix;
 
 #define TILE_SZ 16
-#define TM (TILE_SZ-5)
-#define TN (TILE_SZ-6)
-#define TK (4 * TILE_SZ-8)
+#define TM (TILE_SZ - 4)
+#define TN (TILE_SZ - 4)
+#define TK (4 * TILE_SZ - 16)
 
 #define SG_SZ 16
 
-template <typename T, size_t NUM_ROWS, size_t NUM_COLS> struct big_matrix{
+template <typename T, size_t NUM_ROWS, size_t NUM_COLS> struct big_matrix {
 public:
   T *mat;
 
 public:
   T *get_data() { return mat; }
   void set_data(T *data) { mat = data; }
-  big_matrix(T *data) : mat(data) {
-  }
+  big_matrix(T *data) : mat(data) {}
 };
 
-template <typename T1, typename T2, size_t NUM_ROWS_A, size_t NUM_COLS_A, size_t NUM_ROWS_B,
-          size_t NUM_COLS_B, size_t NUM_ROWS_C, size_t NUM_COLS_C>
-void matrix_multiply(big_matrix<T1, NUM_ROWS_C, NUM_COLS_C> &C, big_matrix<T2, NUM_ROWS_A, NUM_COLS_A> &A, big_matrix<T2, NUM_ROWS_B, NUM_COLS_B> &B) {
+template <typename T1, typename T2, size_t NUM_ROWS_A, size_t NUM_COLS_A,
+          size_t NUM_ROWS_B, size_t NUM_COLS_B, size_t NUM_ROWS_C,
+          size_t NUM_COLS_C>
+void matrix_multiply(big_matrix<T1, NUM_ROWS_C, NUM_COLS_C> &C,
+                     big_matrix<T2, NUM_ROWS_A, NUM_COLS_A> &A,
+                     big_matrix<T2, NUM_ROWS_B, NUM_COLS_B> &B) {
   size_t M = NUM_ROWS_C;
   size_t N = NUM_COLS_C;
   size_t K = NUM_COLS_A;
@@ -46,7 +49,7 @@ void matrix_multiply(big_matrix<T1, NUM_ROWS_C, NUM_COLS_C> &C, big_matrix<T2, N
 
      cgh.parallel_for<class imatrix>(
          nd_range<2>({NDRangeM, NDRangeN * SG_SZ}, {1, 1 * SG_SZ}),
-         [accA, accB, accC, M, N, K](nd_item<2> spmd_item)  [[intel::reqd_sub_group_size(SG_SZ)]]
+         [accA, accB, accC, M, N, K](nd_item<2> spmd_item)
 
          {
            // The submatrix API has to be accessed by all the workitems in a
@@ -59,9 +62,10 @@ void matrix_multiply(big_matrix<T1, NUM_ROWS_C, NUM_COLS_C> &C, big_matrix<T2, N
 
            ext::oneapi::sub_group sg = spmd_item.get_sub_group();
            joint_matrix<int8_t, TM, TK> sub_a(sg);
-           // For B, since current implementation does not support non-packed layout,
-           // users need to specify the updated VNNI sizes along with the packed_b layout.
-           // By default, the layout is row_major and size is (TK, TN).
+           // For B, since current implementation does not support non-packed
+           // layout, users need to specify the updated VNNI sizes along with
+           // the packed_b layout. By default, the layout is row_major and size
+           // is (TK, TN).
            joint_matrix<int8_t, TK, TN, matrix_layout::packed_b> sub_b(sg);
            joint_matrix<int32_t, TM, TN> sub_c(sg);
 
@@ -82,6 +86,10 @@ void matrix_multiply(big_matrix<T1, NUM_ROWS_C, NUM_COLS_C> &C, big_matrix<T2, N
                                N * 4, matrix_layout::packed_b);
              sub_c = joint_matrix_mad(sg, sub_a, sub_b, sub_c);
            }
+           auto wi_data_c = sub_c.get_wi_data();
+           for (int i = 0; i < wi_data_c.length(); i++) {
+             wi_data_c[i] *= 2;
+           }
            joint_matrix_store(sg, sub_c,
                               accC.get_pointer() + (sg_startx * TM) * N +
                                   sg_starty / SG_SZ * TN,
@@ -99,7 +107,7 @@ int32_t C[MATRIX_M][MATRIX_N];
 int32_t D[MATRIX_M][MATRIX_N];
 
 void matrix_multiply_ref(int32_t *A_mem, int32_t *B_mem, int32_t *C_mem, int M,
-                       int N, int K) {
+                         int N, int K) {
   // tiling
   for (int m = 0; m < M; m++)
     for (int n = 0; n < N; n++) {
@@ -112,18 +120,19 @@ void matrix_multiply_ref(int32_t *A_mem, int32_t *B_mem, int32_t *C_mem, int M,
         }
         *(C_mem + m * N + n) = acc;
       }
+      *(C_mem + m * N + n) *= 2;
     }
 }
 
 int main() {
   for (int i = 0; i < MATRIX_M; i++) {
     for (int j = 0; j < MATRIX_K; j++) {
-      A[i][j] = i+2*j;
+      A[i][j] = i + 2 * j;
     }
   }
   for (int i = 0; i < MATRIX_K / 4; i++) {
     for (int j = 0; j < MATRIX_N * 4; j++) {
-      B[i][j] = i+j;
+      B[i][j] = i + j;
     }
   }
   for (int i = 0; i < MATRIX_M; i++) {
@@ -136,10 +145,10 @@ int main() {
   big_matrix<int32_t, MATRIX_M, MATRIX_N> MC((int32_t *)&C);
   big_matrix<int32_t, MATRIX_M, MATRIX_N> MD((int32_t *)&D);
   big_matrix<int8_t, MATRIX_M, MATRIX_K> MA((int8_t *)&A);
-  big_matrix<int8_t,MATRIX_K / 4, MATRIX_N * 4> MB((int8_t *)&B);
+  big_matrix<int8_t, MATRIX_K / 4, MATRIX_N * 4> MB((int8_t *)&B);
   matrix_multiply(MC, MA, MB);
   matrix_multiply_ref((int32_t *)A, (int32_t *)B, (int32_t *)D, MATRIX_M,
-                    MATRIX_N, MATRIX_K / 4);
+                      MATRIX_N, MATRIX_K / 4);
 
   bool res = true;
   for (int i = 0; i < MATRIX_M; i++) {
